Added libnitrokey ebuild.
Libnitrokey must be checked out from the tip of the master branch, as the 2.0
release found in the repo is severely outdated.

Secondly, the cmakelists patch is required in order to allow for the static
libs to be installed and in order to prevent it from linking to the vendored
hidapi and libusb.